requirements.txt
numpy==1.26.4
pandas==2.1.4
beautifulsoup4==4.12.3
requests==2.31.0
openai==1.40.0
python-dotenv==1.0.1
scikit-learn==1.3.2
tqdm==4.66.2
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
